+++ protocol/src/main/kotlin/model/editorPlugin/RdEditorModel.kt
@@ -4,21 +4,18 @@ import com.jetbrains.rd.generator.nova.*
 import com.jetbrains.rd.generator.nova.PredefinedType.*
 import com.jetbrains.rd.generator.nova.cpp.Cpp17Generator
 import com.jetbrains.rd.generator.nova.csharp.CSharp50Generator
-import com.jetbrains.rd.generator.nova.util.syspropertyOrInvalid
 import model.lib.ue4.UE4Library
-import java.io.File
 
 @Suppress("unused")
-object RdEditorRoot : Root(
-        // TODO: use settings for precompiled headers when available and remove hardcoded generator
-        Cpp17Generator(FlowTransform.Reversed, "", File(syspropertyOrInvalid("model.out.src.editorPlugin.cpp.dir")), generatePrecompiledHeaders = false)
-) {
+object RdEditorRoot : Root() {
     init {
         setting(CSharp50Generator.Namespace, "RiderPlugin.UnrealLink.Model.BackendUnreal")
         setting(Cpp17Generator.Namespace, "JetBrains::EditorPlugin")
 
         setting(Cpp17Generator.AdditionalHeaders, listOf("UE4TypesMarshallers.h"))
         setting(Cpp17Generator.ExportMacroName,  "RIDERLINK_API")
+        setting(Cpp17Generator.GeneratePrecompiledHeaders, false)
+        setting(Cpp17Generator.UsePrecompiledHeaders, false)
     }
 }